refactor(openomni): split subagent spawn policy
